+++ apps/client/src/app/core/alarms/alarm-bell.service.ts
@@ -36,7 +36,10 @@ export class AlarmBellService {
             if (alarm.spawns === undefined) {
               return false;
             }
-            const hasOneGroupEnabled = groups.some(group => {
+
+            const groupsForThisAlarm = groups.filter(g => g.alarms.includes(alarm.$key));
+
+            const hasOneGroupEnabled = groupsForThisAlarm.length === 0 || groupsForThisAlarm.some(group => {
               return group.enabled && group.alarms.includes(alarm.$key);
             });
             // If this alarm has a group and it's muted, don't even go further