+++ apps/client/src/app/core/electron/packet-capture-tracker.service.ts
@@ -21,7 +21,7 @@ import { Region } from '../../modules/settings/region.enum';
 import { environment } from '../../../environments/environment';
 import { LazyDataService } from '../data/lazy-data.service';
 import { InventoryEventType } from '../../model/user/inventory/inventory-event-type';
-import { ActorControl, EffectResult, FishingBaitMsg, InitZone, PlayerSetup, UpdateClassInfo, WeatherChange } from '../../model/pcap';
+import { ActorControl, EffectResult, FishingBaitMsg, InitZone, UpdateClassInfo, WeatherChange } from '../../model/pcap';
 
 @Injectable({
   providedIn: 'root'
@@ -143,13 +143,14 @@ export class PacketCaptureTrackerService {
             Object.keys(inventory.items[inventory.contentId])
               .filter(key => key.startsWith(lastRetainerSpawned))
               .forEach(key => inventory.items[inventory.contentId][key] = {});
-          } else {
-            inventory = this.resetInventoryForItemInfo(inventory);
           }
         }
 
         groupedInfos.forEach(group => {
           const containerKey = isRetainer ? `${lastRetainerSpawned}:${group.containerId}` : `${group.containerId}`;
+          if (!isRetainer && +group.containerId % 1000 === 0) {
+            inventory = this.resetInventoryForItemInfo(inventory, +group.containerId);
+          }
           inventory.items[inventory.contentId][containerKey] = {};
           group.packets.forEach(packet => {
             const item: InventoryItem = {
@@ -323,13 +324,23 @@ export class PacketCaptureTrackerService {
         }
       }
     });
+
+    this.ipc.packets$.pipe(
+      ofPacketType('logout')
+    ).subscribe(() => {
+      this.userInventoryService.setContentId(null);
+    });
+
+    this.ipc.playerSetupPackets$.subscribe((packet) => {
+      this.userInventoryService.setContentId(BigInt(packet.contentID).toString(16).padStart(16, '0').toUpperCase());
+    });
   }
 
-  private resetInventoryForItemInfo(inventory: UserInventory): UserInventory {
-    const itemsClone = JSON.parse(JSON.stringify(inventory.items[inventory.contentId]))
+  private resetInventoryForItemInfo(inventory: UserInventory, containerKey: number): UserInventory {
+    const itemsClone = JSON.parse(JSON.stringify(inventory.items[inventory.contentId]));
     Object.keys(itemsClone)
       .forEach(key => {
-        if (key.indexOf(':') === -1 && +key < 4000) {
+        if (key.indexOf(':') === -1 && Math.floor(+key / 1000) === Math.floor(containerKey / 1000)) {
           itemsClone[key] = {};
         }
       });