+++ apps/client/src/app/modules/list/list-details-panel/list-details-panel.component.ts
@@ -35,7 +35,6 @@ import { Vendor } from '../model/vendor';
 import { LayoutRowDisplayMode } from '../../../core/layout/layout-row-display-mode';
 import { NpcBreakdown } from '../../../model/common/npc-breakdown';
 import { NpcBreakdownRow } from '../../../model/common/npc-breakdown-row';
-import { mapIds } from '../../../core/data/sources/map-ids';
 
 @Component({
   selector: 'app-list-details-panel',
@@ -187,7 +186,7 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
           && !this.lazyData.data.maps[d.mapid].dungeon
           && (!zoneId || d.zoneid === zoneId);
       });
-      const hasNodesWithPosition = (getItemSource(row, DataType.GATHERED_BY, true).nodes || []).some(n => n.x !== undefined && (!zoneId || mapIds.find(m => m.id === n.map)?.zone === zoneId));
+      const hasNodesWithPosition = (getItemSource(row, DataType.GATHERED_BY, true).nodes || []).some(n => n.x !== undefined && (!zoneId || n.zoneId === zoneId));
       const hasVendorsWithPosition = getItemSource(row, DataType.VENDORS).some(d => d.coords && (d.coords.x !== undefined) && (!zoneId || d.zoneId === zoneId));
       const hasTradesWithPosition = getItemSource(row, DataType.TRADE_SOURCES).some(d => d.npcs.some(npc => npc.coords && npc.coords.x !== undefined && (!zoneId || npc.zoneId === zoneId)));
       return hasMonstersWithPosition || hasNodesWithPosition || hasVendorsWithPosition || hasTradesWithPosition || hasMap;
@@ -303,7 +302,7 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
       });
     }
     if ((gatheredBy.nodes || []).some(n => n.x !== undefined && n.zoneId === zoneBreakdownRow.zoneId)) {
-      const node = gatheredBy.nodes.find(n => n.x !== undefined && mapIds.find(m => m.id === n.map)?.zone === zoneBreakdownRow.zoneId);
+      const node = gatheredBy.nodes.find(n => n.x !== undefined && n.zoneId === zoneBreakdownRow.zoneId);
       positions.push({
         x: node.x,
         y: node.y,
@@ -436,7 +435,7 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
     if (id === -1) {
       return { fr: 'Autre', de: 'Anderes', ja: 'Other', en: 'Other', zh: '其他', ko: '기타' };
     }
-    return this.l12n.getPlace(id);
+    return this.l12n.getMapName(id);
   }
 
   public getNpc(id: number): I18nName {