+++ apps/client/src/environments/patch-notes.ts
@@ -1,26 +1,11 @@
 export const patchNotes = `### Bug Fixes
 
-* **alarms:** fixed alarms not showing notification on web version.
-* **alarms:** fixed alarms regenerating forever in some cases.
-* **auth:** fixed google login not working properly in some cases.
-* **community-rotations:** fixed an issue with simulation result not showing.
-* **db:** fixed fish page data sometimes not loading properly.
-* **desktop:** fixed various issues with multi character inventory tracking.
-* **layout:** fixed buttons and support menu entries overlapping.
-* **layout:** fixed copy isearch menu entry not working as intended.
-* **leveling-equipment:** fixed wrong suggestions in some cases.
-* **leveling-equipment:** ignore items that are linked to a festival.
-* **list:** fixed path optimizer popups.
-* **nodes:** fixed some nodes showing 4h spawn duration instead of 3h.
-* **retainer-ventures:** fixed result ordering.
-* **retainers:** now showing remaining time instead of end time.
-
-
-### Features
-
-* **db:** added a link from baits to the fish they can catch.
-* **desktop:** new button in settings to unlink your characters from ingame characters.
-* **desktop:** you can now delete one inventory again in inventory page.
-* **loading:** better loading screen for lazy data on web app.
-* **profile:** new entry in top-right profile menu to quickly select another character.
-* **settings:** new button to disconnect from patreon.`;
+* **alarms:** fixed a bug with alarms sometimes not playing the sound notification.
+* **alarms:** fixed an issue with "add alarms to a group" not working properly.
+* **currency-spending:** removed outdated red scrips.
+* **desktop:** better multi character detection combining packets and proper content checks.
+* **desktop:** fixed a bug with inventory resetting in some cases.
+* **list:** fixed requirements popup not showing on website.
+* **list:** fixed wrong tooltip on copy menu entry.
+* **list:** fixed zone breakdown sometimes not loading properly.
+* **retainer-ventures:** fixed retainer selector not showing properly.`;